requirements.txt
cachetools
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-telegram-bot[job-queue]==20.7
PyYAML
requests_oauthlib
typing_extensions
tweepy==4.14.0
